GUI Screens

+++ src/main/java/edu/ncsu/csc573/project/common/messages/IStatus.java
@@ -1,4 +1,7 @@
 package edu.ncsu.csc573.project.common.messages;
+
+import java.math.BigInteger;
+
 /**
  * This interface abstracts Status component in Response.
  * @author doogle-dev
@@ -15,6 +18,6 @@ public interface IStatus {
 	 * return error id if there is Response has an error.
 	 * @return
 	 */
-	public int getErrorId();
+	public BigInteger getErrorId();
 	
 }
